// src/apis/models/FireBytes_Unified_Wave_Pacs_PacsPublicHolidays_Dtos_PacsPublicHolidayEditDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_Pacs_PacsPublicHolidays_Dtos_PacsPublicHolidayEditDto = {
    id?: string | null;
    holidayName?: string | null;
    isEveOfHoliday?: boolean;
    startDate?: string;
    endDate?: string;
};